+++ jeap-error-handling-service/src/test/java/ch/admin/bit/jeap/errorhandling/domain/error/SyncWithManualTasksTest.java
@@ -7,7 +7,7 @@ import ch.admin.bit.jeap.errorhandling.domain.manualtask.taskFactory.TaskFactory
 import ch.admin.bit.jeap.errorhandling.domain.metrics.ErrorHandlingMetricsService;
 import ch.admin.bit.jeap.errorhandling.domain.resend.scheduler.ScheduledResendService;
 import ch.admin.bit.jeap.errorhandling.domain.resend.strategy.ResendingStrategy;
-import ch.admin.bit.jeap.errorhandling.infrastructure.kafka.KafkaFailedEventResender;
+import ch.admin.bit.jeap.errorhandling.infrastructure.kafka.FailedEventResender;
 import ch.admin.bit.jeap.errorhandling.infrastructure.manualtask.TaskDto;
 import ch.admin.bit.jeap.errorhandling.infrastructure.manualtask.TaskManagementClient;
 import ch.admin.bit.jeap.errorhandling.infrastructure.manualtask.TaskManagementException;
@@ -48,7 +48,7 @@ class SyncWithManualTasksTest {
     @MockitoBean
     private ScheduledResendService scheduledResendService;
     @MockitoBean
-    private KafkaFailedEventResender failedEventResender;
+    private FailedEventResender failedEventResender;
     @MockitoBean
     private TaskManagementClient taskManagementClient;
     @MockitoBean