+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [22.6.0] - 2026-09-01
+
+### Added
+- Handle failed Spring Modulith publications and send retry or discard commands through the transactional outbox.
+- Show the error origin and Modulith publication details and actions in the UI.
+
 ## [22.5.1] - 2026-08-31
 
 ### Dependencies